requirements.txt
channels==2.1.3
daphne==2.2.2
Django==2.1.2
django-polymorphic==2.0.3
python-magic==0.4.15
Pillow==5.3.0
django-bootstrap3==11.0.0
gunicorn==19.9.0
psycopg2==2.7.5
whitenoise==4.1
dj-database-url==0.4.1
channels-redis==2.3.0
djangorestframework==3.8.2
django-rest-knox==3.1.5